added support for xdg-open
space fix

+++ graphviz.sh
@@ -2,6 +2,11 @@ INPUT_FILE=$1
 DOT_FILE=$INPUT_FILE.dot
 PNG_FILE=$INPUT_FILE.png
 
+if [ "$INPUT_FILE" = "" ]; then
+	echo "usage: ./graphviz.sh input_file"
+	exit
+fi
+
 rm -rf "$PNG_FILE" "$DOT_FILE"
 
 # dot file generation
@@ -24,4 +29,11 @@ echo "}" >> "$DOT_FILE"
 # rendering
 dot -Tpng "$DOT_FILE" > "$PNG_FILE"
 
-gnome-open "$PNG_FILE"
+# open the png file
+if command -v xdg-open &> /dev/null
+then
+	xdg-open "$PNG_FILE"
+elif command -v gnome-open &> /dev/null
+then
+	gnome-open "$PNG_FILE"
+fi
